GameRules class added (stubs)
I'm a bit unsure on what should be included here after I saw the
GameState class. Maybe move the CurrentPlayer to GameState, and bring
methods like isDone() and Winner() into GameRules? Or merge them both
into one class?

GameState should probably deal with stuff like how the subgames are tied
together into the larger game (supergame?) and which subgame/square was
selected by the player. So it passes in the essential information to be
crunched by GameRules, and then makes sense of the returned values in
order to store them for the UI guys to access. Yes/No/Confused?